+++ packages/backend/package.json
@@ -46,6 +46,7 @@
   },
   "dependencies": {
     "cors": "^2.8.5",
+    "express": "^4.18.2",
     "jsonwebtoken": "^8.5.1",
     "koa": "^2.14.1",
     "koa-body": "^4.2.0",
@@ -56,6 +57,7 @@
     "koa-static": "^5.0.0",
     "md5": "^2.3.0",
     "mongoose": "^5.11.4",
+    "public-ip": "^6.0.1",
     "socket.io": "^4.5.4",
     "ws": "^8.11.0"
   }
